requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# Python 3.7
# Original code in Pytorch 1.0.0
# Compatibility patch for Pytorch 1.6 (Create git issue for remaining bugs)
_libgcc_mutex=0.1=main
blas=1.0=mkl
ca-certificates=2020.6.20=hecda079_0
certifi=2020.6.20=py37hc8dfbb8_0
chardet=3.0.4=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py_2
freetype=2.10.2=h5ab3b9f_0
future=0.18.2=pypi_0
idna=2.10=pypi_0
intel-openmp=2020.2=254
jpeg=9b=h024ee3a_2
jsonpatch=1.26=pypi_0
jsonpointer=2.0=pypi_0
kiwisolver=1.2.0=py37h99015e2_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
lz4-c=1.9.2=he6710b0_1
matplotlib=3.3.1=1
matplotlib-base=3.3.1=py37hd478181_1
mkl=2020.2=256
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.1.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
ncurses=6.2=he6710b0_1
ninja=1.10.1=py37hfd86e86_0
numpy=1.19.1=py37hbc911f0_0
numpy-base=1.19.1=py37hfa32c7d_0
olefile=0.46=py37_0
openssl=1.1.1g=h516909a_1
pillow=7.2.0=py37hb39fc2d_0
pip=20.2.2=py37_0
pyparsing=2.4.7=pyh9f0ad1d_0
python=3.7.9=h7579374_0
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.6.0=py3.7_cuda10.2.89_cudnn7.6.5_0
pyzmq=19.0.2=pypi_0
quadprog=0.1.6=py37_0
readline=8.0=h7b6447c_0
requests=2.24.0=pypi_0
scipy=1.5.2=pypi_0
setuptools=49.6.0=py37_0
six=1.15.0=py_0
sqlite=3.33.0=h62c20be_0
tk=8.6.10=hbc83047_0
torchfile=0.1.0=pypi_0
torchnet=0.0.5.1=pypi_0
torchvision=0.7.0=py37_cu102
tornado=6.0.4=py37h8f50634_1
tqdm=4.48.2=pyh9f0ad1d_0
urllib3=1.25.10=pypi_0
visdom=0.1.8.9=pypi_0
websocket-client=0.57.0=pypi_0
wheel=0.35.1=py_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
